refactor: 콘솔앱의 main 파일 함수들을 BankManager 모듈의 Bank로 이동
